+++ server/consul.go
@@ -15,6 +15,10 @@
 package server
 
 import (
+	"io"
+	"os"
+	"strconv"
+
 	"github.com/blang/semver"
 	"github.com/hashicorp/consul/api"
 	"github.com/pkg/errors"
@@ -31,12 +35,22 @@ var upgradeToMap = map[string]func(*api.KV, <-chan struct{}) error{
 }
 
 var orderedUpgradesVersions []semver.Version
+var disableConsulSnapshotsOnUpgrades bool
 
 func init() {
 	for v := range upgradeToMap {
 		orderedUpgradesVersions = append(orderedUpgradesVersions, semver.MustParse(v))
 	}
 	semver.Sort(orderedUpgradesVersions)
+
+	var err error
+	disableConsulSnapEnv := os.Getenv("YORC_DISABLE_CONSUL_SNAPSHOTS_ON_UPGRADE")
+	if disableConsulSnapEnv != "" {
+		disableConsulSnapshotsOnUpgrades, err = strconv.ParseBool(disableConsulSnapEnv)
+		if err != nil {
+			log.Panicf(`Can't read value of "YORC_DISABLE_CONSUL_SNAPSHOTS_ON_UPGRADE" environment variable: %v`, err)
+		}
+	}
 }
 
 func synchronizeDBUpdate(client *api.Client) (*api.Lock, <-chan struct{}, error) {
@@ -121,32 +135,45 @@ func upgradeFromVersion(client *api.Client, leaderCh <-chan struct{}, fromVersio
 		// Same version nothing to do
 		return nil
 	case 1:
+		err = performUpgrade(client, leaderCh, vCurrent)
+		if err != nil {
+			return err
+		}
+	case -1:
+		return errors.Errorf("this version of Yorc is too old compared to the current DB schema (%s), an upgrade is needed.", vCurrent)
+	}
+
+	return setNewVersion(client.KV())
+}
+
+func performUpgrade(client *api.Client, leaderCh <-chan struct{}, vCurrent semver.Version) error {
+	snap := client.Snapshot()
+	var snapReader io.ReadCloser
+	if !disableConsulSnapshotsOnUpgrades {
 		// Make a Consul snapshot and restore it if any error occurs
-		snap := client.Snapshot()
-		snapReader, _, err := snap.Save(nil)
+		var err error
+		snapReader, _, err = snap.Save(nil)
 		if err != nil {
 			return errors.Wrapf(err, "failed to upgrade consul db schema to %q", err)
 		}
 		defer snapReader.Close()
-		for _, vUp := range orderedUpgradesVersions {
-			if vUp.GT(vCurrent) {
-				err = upgradeToMap[vUp.String()](client.KV(), leaderCh)
-				if err != nil {
+	}
+	for _, vUp := range orderedUpgradesVersions {
+		if vUp.GT(vCurrent) {
+			err := upgradeToMap[vUp.String()](client.KV(), leaderCh)
+			if err != nil {
+				if !disableConsulSnapshotsOnUpgrades {
 					// Restore Consul snapshot
 					restoreErr := snap.Restore(nil, snapReader)
 					if restoreErr != nil {
-						log.Printf("failed to restore consul db schema to %q due to error:%+v", fromVersion, restoreErr)
+						log.Printf("failed to restore consul db schema to %q due to error:%+v", vCurrent, restoreErr)
 					} else {
-						log.Printf("As any error occurred, schema has been successfully restored to version %q", fromVersion)
+						log.Printf("As any error occurred, schema has been successfully restored to version %q", vCurrent)
 					}
-					return errors.Wrapf(err, "failed to upgrade consul db schema to %q.", vUp)
 				}
+				return errors.Wrapf(err, "failed to upgrade consul db schema to %q.", vUp)
 			}
 		}
-	case -1:
-		return errors.Errorf("this version of Yorc is too old compared to the current DB schema (%s), an upgrade is needed.", vCurrent)
-
 	}
-
-	return setNewVersion(client.KV())
+	return nil
 }