+++ tasks/collector/collector.go
@@ -19,7 +19,6 @@ import (
 	"fmt"
 	"path"
 	"strconv"
-	"strings"
 	"time"
 
 	"github.com/hashicorp/consul/api"
@@ -219,16 +218,10 @@ func (c *Collector) prepareForRegistration(operations api.KVTxnOps, taskType tas
 		})
 	}
 
-	ok, response, _, err := c.consulClient.KV().Txn(operations, nil)
+	err := tasks.StoreOperations(c.consulClient.KV(), taskID, operations)
 	if err != nil {
-		return errors.Wrapf(err, "Failed to register task with targetID:%q, taskType:%q", targetID, taskType.String())
-	}
-	if !ok {
-		errs := make([]string, 0)
-		for _, e := range response.Errors {
-			errs = append(errs, e.What)
-		}
-		return errors.Wrapf(err, "Failed to register task with targetID:%q, taskType:%q due to error:%s", targetID, taskType.String(), strings.Join(errs, ", "))
+		return errors.Wrapf(err, "Failed to register task with targetID:%q, taskType:%q due to error %s",
+			targetID, taskType.String(), err.Error())
 	}
 
 	log.Debugf("Registered task %s type %q for target %s\n", taskID, taskType.String(), targetID)