+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Yorc Changelog
 
+## UNRELEASED
+
 ## 3.2.3 (July 16, 2019)
 
 ### ENHANCEMENTS